home/user/priyambodo.com-iamagentic-adk/project21-adk-withcheatcode/iam_stock_agent/requirements.txt
google-adk
yfinance
